requirements.txt
langchain==0.3.25
langchain_anthropic==0.3.13
matplotlib==3.10.3
numpy==2.2.5
pandas==2.2.3
protobuf==6.31.0
pydantic==2.11.4
python-dotenv==1.1.0
seaborn==0.13.2
torch==2.2.2
transformers==4.51.3
